Rename hooks/auth to gitprovider.ts
